mods filesystem client tests so they can be run separately
